requirements.txt
numpy==1.22.4
pandas==1.4.3
scikit_learn==1.3.0
matplotlib==3.5.2
networkx==3.1
PyYAML==6.0.1
scipy==1.8.1
tensorboardX==2.6.2.2
torch==1.11.0
torchvision==0.12.0
tqdm==4.63.0